refactor DigestStream
